+++ cmd/template/networkpool/flag.go
@@ -12,7 +12,8 @@ const (
 	flagCIDRBlock       = "cidr-block"
 	flagNetworkPoolName = "networkpool-name"
 	flagOutput          = "output"
-	flagOwner           = "owner"
+	flagOrganization    = "org"
+	flagOwner           = "owner" // TODO: Remove some time after December 2021
 )
 
 type flag struct {
@@ -20,6 +21,7 @@ type flag struct {
 	CIDRBlock       string
 	NetworkPoolName string
 	Output          string
+	Organization    string
 	Owner           string
 }
 
@@ -27,7 +29,11 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.NetworkPoolName, flagNetworkPoolName, "", "NetworkPool identifier.")
 	cmd.Flags().StringVar(&f.CIDRBlock, flagCIDRBlock, "", "Installation infrastructure provider.")
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", "File path for storing CRs. (default: stdout)")
+	cmd.Flags().StringVar(&f.Organization, flagOrganization, "", "Workload cluster organization.")
 	cmd.Flags().StringVar(&f.Owner, flagOwner, "", "Workload cluster owner organization.")
+
+	// TODO: Remove around December 2021
+	_ = cmd.Flags().MarkDeprecated(flagOwner, "please use --org instead.")
 }
 
 func (f *flag) Validate() error {
@@ -42,8 +48,19 @@ func (f *flag) Validate() error {
 			return microerror.Maskf(invalidFlagError, "--%s must be a valid CIDR", flagCIDRBlock)
 		}
 	}
-	if f.Owner == "" {
-		return microerror.Maskf(invalidFlagError, "--%s must not be empty", flagOwner)
+
+	// TODO: Remove the flag completely some time after December 2021
+	if f.Owner == "" && f.Organization == "" {
+		return microerror.Maskf(invalidFlagError, "--%s or --%s must not be empty", flagOwner, flagOrganization)
+	}
+
+	if f.Owner != "" {
+		if f.Organization != "" {
+			return microerror.Maskf(invalidFlagError, "--%s and --%s cannot be combined", flagOwner, flagOrganization)
+		}
+
+		f.Organization = f.Owner
+		f.Owner = ""
 	}
 
 	return nil