+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- In the `template` commands, the flag `--owner` is replaced by `--organization`.
 - Make the `login` command be able to start a new authentication flow if one of the tokens of an existing authentication provider are not present.
 
 ## [1.39.0] - 2021-09-10